refactor: [M3-9387] - Move Stackscripts queries
