+++ packages/manager/src/features/Linodes/LinodeCreate/Tabs/StackScripts/StackScriptDetailsDialog.tsx
@@ -1,8 +1,9 @@
+// import { useStackScriptQuery } from 'src/queries/stackscripts';
+import { useStackScriptQuery } from '@linode/queries';
 import { CircleProgress, Dialog, ErrorState } from '@linode/ui';
 import React from 'react';
 
 import { StackScript } from 'src/components/StackScript/StackScript';
-import { useStackScriptQuery } from 'src/queries/stackscripts';
 
 interface Props {
   /**
