+++ packages/manager/src/features/Linodes/LinodeCreate/Tabs/StackScripts/UserDefinedFields/UserDefinedFields.tsx
@@ -1,3 +1,4 @@
+import { useStackScriptQuery } from '@linode/queries';
 import { Box, IconButton, Notice, Paper, Stack, Typography } from '@linode/ui';
 import React from 'react';
 import { useFormContext, useWatch } from 'react-hook-form';
@@ -5,7 +6,6 @@ import { useFormContext, useWatch } from 'react-hook-form';
 import Info from 'src/assets/icons/info.svg';
 import { ShowMoreExpansion } from 'src/components/ShowMoreExpansion';
 import { oneClickApps } from 'src/features/OneClickApps/oneClickApps';
-import { useStackScriptQuery } from 'src/queries/stackscripts';
 
 import { getMarketplaceAppLabel } from '../../Marketplace/utilities';
 import { UserDefinedFieldInput } from './UserDefinedFieldInput';
