+++ packages/manager/src/features/Linodes/LinodeCreate/utilities.ts
@@ -1,4 +1,8 @@
-import { accountQueries, linodeQueries } from '@linode/queries';
+import {
+  accountQueries,
+  linodeQueries,
+  stackscriptQueries,
+} from '@linode/queries';
 import { omitProps } from '@linode/ui';
 import {
   getQueryParamsFromQueryString,
@@ -10,7 +14,6 @@ import { useCallback } from 'react';
 import { useHistory } from 'react-router-dom';
 
 import { imageQueries } from 'src/queries/images';
-import { stackscriptQueries } from 'src/queries/stackscripts';
 import { sendCreateLinodeEvent } from 'src/utilities/analytics/customEventAnalytics';
 import { sendLinodeCreateFormErrorEvent } from 'src/utilities/analytics/formEventAnalytics';
 import { isPrivateIP } from 'src/utilities/ipUtils';
