+++ packages/manager/src/features/Linodes/LinodesDetail/LinodeRebuild/UserDefinedFields.tsx
@@ -1,9 +1,9 @@
+import { useStackScriptQuery } from '@linode/queries';
 import { Box, Notice, Stack, Typography } from '@linode/ui';
 import React from 'react';
 import { useFormContext, useWatch } from 'react-hook-form';
 
 import { ShowMoreExpansion } from 'src/components/ShowMoreExpansion';
-import { useStackScriptQuery } from 'src/queries/stackscripts';
 
 import { UserDefinedFieldInput } from '../../LinodeCreate/Tabs/StackScripts/UserDefinedFields/UserDefinedFieldInput';
 import { separateUDFsByRequiredStatus } from '../../LinodeCreate/Tabs/StackScripts/UserDefinedFields/utilities';
