+++ packages/manager/src/features/Search/useClientSideSearch.ts
@@ -2,6 +2,7 @@ import { useAllLinodesQuery } from '@linode/queries';
 import { useAllFirewallsQuery } from '@linode/queries';
 import { useAllVolumesQuery } from '@linode/queries';
 import { useAllNodeBalancersQuery } from '@linode/queries';
+import { useAllAccountStackScriptsQuery } from '@linode/queries';
 
 import { useKubernetesBetaEndpoint } from 'src/features/Kubernetes/kubeUtils';
 import { useAllDatabasesQuery } from 'src/queries/databases/databases';
@@ -9,7 +10,6 @@ import { useAllDomainsQuery } from 'src/queries/domains';
 import { useAllImagesQuery } from 'src/queries/images';
 import { useAllKubernetesClustersQuery } from 'src/queries/kubernetes';
 import { useObjectStorageBuckets } from 'src/queries/object-storage/queries';
-import { useAllAccountStackScriptsQuery } from 'src/queries/stackscripts';
 import {
   bucketToSearchableItem,
   databaseToSearchableItem,
