+++ packages/manager/src/features/StackScripts/stackScriptUtils.ts
@@ -1,28 +1,4 @@
-import { getStackScripts } from '@linode/api-v4';
-
-import type {
-  Grant,
-  Params,
-  StackScript,
-  StackScriptPayload,
-} from '@linode/api-v4';
-
-const oneClickFilter = [
-  {
-    '+and': [
-      { '+or': [{ username: 'linode-stackscripts' }, { username: 'linode' }] },
-      {
-        label: {
-          '+contains': 'One-Click',
-        },
-      },
-    ],
-    '+order_by': 'ordinal',
-  },
-];
-
-export const getOneClickApps = (params?: Params) =>
-  getStackScripts(params, oneClickFilter);
+import type { Grant, StackScript, StackScriptPayload } from '@linode/api-v4';
 
 export const getStackScriptUrl = (
   username: string,
